wayland: refactor break out layerstatic,subsurface to layer.py
